Renamed GridCardContainer to GridContainer since it can be used for more than just cards
